refactor: move calendar to main gui file to get rid of helper file
